added / fixed tests
