+++ react_app/src/Components/Clock.js
@@ -1,10 +1,16 @@
 import { useState, useEffect } from 'react';
 
+
+
 const Clock = () => {
   const [date, setDate] = useState(new Date());
   
+  function refreshClock(){
+    setDate(new Date());
+  }
+
   useEffect(() => {
-    const timerId = setInterval(setDate(new Date()), 1000);
+    const timerId = setInterval(refreshClock, 1000);
     return function cleanup() {
       clearInterval(timerId);
     };
